Corrección de toma de segundo,
antes no tomaba el registro números decimales, ahora si. se ve reflejado

+++ app/src/main/java/com/example/pitstop/modelo/Parada.kt
@@ -3,7 +3,7 @@ package com.example.pitstop.modelo
 data class Parada(
     val piloto: Piloto,
     val escuderia: Escuderia,
-    val tiempoSegundos: Int,
+    val tiempoSegundos: Double,
     val neumatico: Neumatico,
     val estado: Estado,
     val observacion: String?,       // puede ser nula
